# projet dash/herbute/backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: ReclamTrack backend API
  description: Backend API for ReclamTrack - IT and Complaint Management System
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StandardResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        timestamp:
          type: string
          format: date-time
          example: '2026-02-17T21:18:22+01:00'
    StandardError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: 'Resource not found'
        code:
          type: string
          example: 'NOT_FOUND'
        timestamp:
          type: string
          format: date-time
        details:
          type: object

    # --- Auth Models ---
    User:
      type: object
      properties:
        id:
          type: string
          description: User ID
        email:
          type: string
          format: email
        role:
          type: string
          enum: [USER, ADMIN, OWNER, TECH_LEAD, TECHNIQUE]
          default: USER
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6

    # --- Complaint Models ---
...
  responses:
...
    UnauthorizedError:
      description: Unauthorized Access
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/StandardError'
              - properties:
                  code:
                    example: UNAUTHORIZED
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - properties:
                      data:
                        type: object
                        properties:
                          token:
                            type: string
                          user:
                            $ref: '#/components/schemas/User'
        '409':
          description: Email already in use
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - properties:
                      data:
                        type: object
                        properties:
                          token:
                            type: string
                          user:
                            $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh authentication token
      security:
        - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - properties:
                      data:
                        type: object
                        properties:
                          token:
                            type: string
                          user:
                            $ref: '#/components/schemas/User'
...